+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-04-19
 ### Added
 
 - `services.addAuthWhitelistedIp()` resource implementation with tests and usage example